@@ -99,8 +99,8 @@
               <groupId>org.apache.maven.plugins</groupId>
               <artifactId>maven-compiler-plugin</artifactId>
               <configuration>
-                  <source>8</source>
-                  <target>8</target>
+                  <source>10</source>
+                  <target>10</target>
               </configuration>
           </plugin>
       </plugins>
