@@ -6,10 +6,10 @@ import java.io.IOException;
 import java.util.List;
 import java.util.Optional;
 
-public interface ClienteDao {
+public interface LaudoDao {
+
     Cliente inserirArquivo(Cliente cliente) throws IOException;
-
     List<Cliente> getAll() throws IOException;
-
     Optional<Cliente> findByCpf(String cpf) throws IOException;
+
 }
