@@ -6,7 +6,10 @@ import lombok.Setter;
 @Getter
 @Setter
 public class Exame {
-    private String nome;
-    private String id;
+
+    private String nomeExame;
+    private String idExame;
     private String parametros;
+    private Cliente cliente;
+
 }
